requirements.txt
blinker==1.6.3
cachelib==0.10.2
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
cs50==9.3.0
Flask==3.0.0
Flask-JWT-Extended==4.5.3
Flask-Session==0.5.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
SQLAlchemy==2.0.22
sqlparse==0.4.4
termcolor==2.3.0
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.1
